==> VU_APRIL_tesinys/0_zive_collect_records_from_list.ipynb <==
# Skriptas nuskaito iš sąrašo nurodytus Zive įrašų failus ir perrašo juos į užduotą aplanką

import os, shutil, sys
from pathlib import Path
import pandas as pd
import numpy as np

import json
from zive_util_vu import create_dir

import warnings
# warnings.filterwarnings("ignore")

my_os=sys.platform
print("OS in my system : ",my_os)

if my_os != 'linux':
    OS = 'Windows'
else:  
    OS = 'Ubuntu'

# Bendras duomenų aplankas, kuriame patalpintas subfolderis name_db

if OS == 'Windows':
    Duomenu_aplankas = 'D:\\DI'   # variantas: Windows
else:
    Duomenu_aplankas = '/home/kesju/DI'   # arba variantas: UBUNTU, be Docker

# jei variantas Docker pasirenkame:
# Duomenu_aplankas = '/Data/MIT&ZIVE'

# Vietinės talpyklos aplankas
db_folder = 'DUOM_2022_RUDUO_3'

# Nuoroda į aplanką su EKG duomenų rinkiniu
db_path = Path(Duomenu_aplankas, db_folder, 'records_selected') # Aplankas, iš kurio imami duomenys

# Nuoroda į aplanką su atrinktų EKG duomenų rinkiniu
db_path_selected = Path(Duomenu_aplankas, db_folder, '20_records_for_testing') # Aplankas, į kurį rašoma

# VARIANTAS, KAI UŽDUODAMI FAILŲ VARDAI IR SKAITOMI ORIGINALUS ZIVE FAILAI 
# 20 atrinktų testinių sąrašas

FileNames = [
1626934.963,
1626931.201,
1630715.664,
1630714.569,
1630729.576,
1630735.143,
1630693.635,
1630734.526,
1630718.396,
1630721.49,
1631139.883,
1631083.411,
1631039.923,
1631029.786,
1632342.032,
1633428.56,
1633584.898,
1633405.853,
1634112.089,
1636451.86
]

# ////////////////////////////////////////////////////////////////////////////////////

# Sukūriame aplanką EKG sekų įrašymui
create_dir(db_path_selected)

# Failų vardus konvertuojame į tekstinius
file_names=["%.3f" % i for i in FileNames]

print("\nSkriptas zive įrašų atrankai ir įrašymui į užduotą aplanką\n")
print( 'Duomenų aplankas skaitymui:', db_path)
print( 'Duomenų aplankas rašymui:', db_path_selected)
print("Atrenkamų failų sąrašas:", file_names)

# Ciklas per visą įrašų sąrašą
for idx, file_name in enumerate(file_names):
    print(f"\nZive įrašas:  {file_name:>2}")

    src = Path(db_path, file_name)
    dst = Path(db_path_selected, file_name)

    # kopijuojame duomenis ir .json failus, išsaugodami metaduomenis su copy2
    shutil.copy2(src, dst)
    src = Path(db_path, file_name + '.json')
    dst = Path(db_path_selected, file_name + '.json')
    shutil.copy2(src, dst)
